Main scene updated with movement and pickup.
Forklift rudimentary movement in place with control to lift and lower Pickup objects on the Pickup LayerMask. Added DOTween implementation.
Current tasks: Improve forklift movement, tweak the max and min lift height to match better with model, resolve issue of controls not working for a few seconds at a time as the system checks for if the lift is animating.